# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tournament Management System",
        "description": "An application to manage tournament, matches and players.",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/matches": {
            "get": {
                "tags": [
                    "Matches"
                ],
                "summary": "Get a list of all tournament matches with associated players",
                "operationId": "7faa70791c8180d28f226f30b05f7fc1",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer (JWT) token for authentication",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved a list of matches",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "matches": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "1"
                                                    },
                                                    "tournament_id": {
                                                        "type": "string",
                                                        "example": "1"
                                                    },
                                                    "date": {
                                                        "type": "string",
                                                        "example": "2025-04-01T10:00:00"
                                                    },
                                                    "location": {
                                                        "type": "string",
                                                        "example": "Stadium A"
                                                    },
                                                    "players": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "string",
                                                                    "example": "10"
                                                                },
                                                                "first_name": {
                                                                    "type": "string",
                                                                    "example": "John"
                                                                },
                                                                "last_name": {
                                                                    "type": "string",
                                                                    "example": "Doe"
                                                                },
                                                                "number": {
                                                                    "type": "integer",
                                                                    "example": 10
                                                                },
                                                                "nationality": {
                                                                    "type": "string",
                                                                    "example": "American"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Failed to retrieve matches"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Matches"
                ],
                "summary": "Create a new tournament match",
                "operationId": "00c9dac2f04ccdce09e29aa1e9695d44",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer (JWT) token for authentication",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Match created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Created Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "date": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The date field is required."
                                                    }
                                                },
                                                "time": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The time field is required."
                                                    }
                                                },
                                                "tournament_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The tournament id field is required."
                                                    }
                                                },
                                                "city": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The city field is required."
                                                    }
                                                },
                                                "host_team_name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The host team name field is required."
                                                    }
                                                },
                                                "guest_team_name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The guest team name field is required."
                                                    }
                                                },
                                                "players": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The players field is required."
                                                    }
                                                },
                                                "host_team_score": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The host team score field is required."
                                                    }
                                                },
                                                "guest_team_score": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The guest team score field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Failed to create match"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}